refactor controller files
