+++ server/src/controllers/matches/documentUserSwipeAccept/index.js
@@ -2,9 +2,9 @@ const MatchRecords = require("../../../models/MatchRecords")
 const MatchQueue = require("../../../models/MatchQueue")
 const { DATA_KEYS } = require("../../../../config/constants")
 
-const fetchMatchRecord = require("../../../utils/matches/fetchMatchRecord")
-const updateUserMatchRecord = require("../../../utils/matches/updateUserMatchRecord")
-const fetchUserMatchQueue = require("../../../utils/matches/fetchUserMatchQueue")
+const fetchMatchRecord = require("../../controllerHelpers/matches/fetchMatchRecord")
+const updateUserMatchRecord = require("../../controllerHelpers/matches/updateUserMatchRecord")
+const fetchUserMatchQueue = require("../../controllerHelpers/matches/fetchUserMatchQueue")
 const verifyEndpointResponse = require("../../../utils/verifyEndpointResponse")
 
 const handleMutualAccept = require("./handleMutualAccept")
