+++ server/src/controllers/zipcodes/index.js
@@ -2,10 +2,10 @@ const ZipcodePool = require("../../models/ZipcodePool")
 const { DATA_KEYS } = require("../../../config/constants")
 const verifyEndpointRequest = require("../../utils/verifyEndpointRequest")
 const verifyEndpointResponse = require("../../utils/verifyEndpointResponse")
-const updateUserDocument = require("../../utils/user/updateUserDocument")
-const fetchUserDocument = require("../../utils/user/fetchUserDocument")
-const addUserToZipcodePool = require("../../utils/zipcode/addUserToZipcodePool")
-const removeUserFromZipcodePool = require("../../utils/zipcode/removeUserFromZipcodePool")
+const updateUserDocument = require("../controllerHelpers/user/updateUserDocument")
+const fetchUserDocument = require("../controllerHelpers/user/fetchUserDocument")
+const addUserToZipcodePool = require("../controllerHelpers/zipcodes/addUserToZipcodePool")
+const removeUserFromZipcodePool = require("../controllerHelpers/zipcodes/removeUserFromZipcodePool")
 const handleErrorResponse = require("../../utils/handleErrorResponse")
 
 const idKey = DATA_KEYS["USER_ID"]
